agent-py/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "SergioCantera",email = "sgfcantera@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "python-dotenv (>=1.2.1,<2.0.0)",
    "langgraph (>=1.0.2,<2.0.0)",
    "langchain-openai (>=1.0.2,<2.0.0)",
    "langchain (>=1.0.5,<2.0.0)",
    "fastapi (>=0.115.0,<0.122.0)",
    "uvicorn (>=0.38.0,<0.39.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "app.api:main"
